+++ primitives/src/merkle_tree/sparse_merkle_tree.rs
@@ -5,7 +5,34 @@
 // along with the Jellyfish library. If not, see <https://mit-license.org/>.
 
 //! Implementation of a typical Sparse Merkle Tree.
+use super::{
+    internal::{build_tree_internal, MerkleNode, MerkleProof},
+    DigestAlgorithm, Element, Index, LookupResult, MerkleCommitment, MerkleTreeScheme, NodeValue,
+    UniversalMerkleTreeScheme,
+};
+use crate::{errors::PrimitivesError, generate_merkle_tree_scheme};
+use ark_std::{borrow::Borrow, boxed::Box, fmt::Debug, marker::PhantomData, string::ToString};
+use num_bigint::BigUint;
+use num_traits::pow::pow;
+use serde::{Deserialize, Serialize};
+use typenum::Unsigned;
 
-// TODO(Chengyu): place holder for future implementation.
-/// a generic SMT
-pub struct SparseMerkleTree;
+// A standard Universal Merkle tree implementation
+generate_merkle_tree_scheme!(UniversalMerkleTree);
+
+impl<E, H, I, Arity, T> UniversalMerkleTreeScheme for UniversalMerkleTree<E, H, I, Arity, T>
+where
+    E: Element,
+    H: DigestAlgorithm<E, I, T>,
+    I: Index + From<u64>,
+    Arity: Unsigned,
+    T: NodeValue,
+{
+    type NonMembershipProof = ();
+
+    type BatchNonMembershipProof = ();
+
+    fn update(&mut self, _pos: Self::Index, _elem: &Self::Element) -> Result<(), PrimitivesError> {
+        todo!()
+    }
+}
